+++ NEWS.md
@@ -1,3 +1,19 @@
+# AMR 1.0.0.9000
+## <small>Last updated: 20-Feb-2020</small>
+
+### Changed
+* Added antibiotic abbreviations for a laboratory manufacturer (GLIMS) for cefuroxime, cefotaxime, ceftazidime, cefepime, cefoxitin and trimethoprim/sulfamethoxazole
+* Fixed floating point error for some MIC compa in EUCAST 2020 guideline
+* Interpretation from MIC values to R/SI can now be used with `mutate_at()` of the dplyr package:
+  ```r
+  yourdata %>% 
+    mutate_at(vars(antibiotic1:antibiotic25), as.rsi, mo = "E. coli")
+    
+  yourdata %>% 
+    mutate_at(vars(antibiotic1:antibiotic25), as.rsi, mo = .$mybacteria)
+  ```
+* Added `uti` (as abbreviation of urinary tract infections) as parameter to `as.rsi()`, so interpretation of MIC values and disk zones can be made dependent on isolates specifically from UTIs
+
 # AMR 1.0.0
 
 This software is now out of beta and considered stable. Nonetheless, this package will be developed continually.